# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CeetFeedback API",
        "description": "API documentation for CeetFeedback",
        "contact": {
            "email": "blezanatanael@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "V1"
    },
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un nouvel utilisateur",
                "operationId": "2b9439118bdeb18e4d0bc425d414fe6e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Connexion d'un utilisateur",
                "operationId": "b2629c0605441a4d4f393a56419bc654",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "StrongPass123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "operationId": "ea36ce7803be314dc7e272620079f305",
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserSchema": {
                "title": "Utilisateur",
                "description": "Ressource utilisateur",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "rule": {
                        "type": "string",
                        "example": "user"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}